refactor: extract sync/snapshots/ + sync/pruner.py → snapshots/
Extract snapshot operations and pruning into a top-level snapshots/
package:

- sync/snapshots/ → snapshots/ (btrfs, hardlinks, common)
- sync/pruner.py → snapshots/pruner.py
- PruneResult model → snapshots/models.py (was in sync/runner.py)
- Remove PruneResult re-export from sync/__init__.py

Dependencies: snapshots → config, fsprotocol, remote, preflight.
No circular deps — sync depends on snapshots, not the reverse.
