+++ nbkp/cli/prune_cmd.py
@@ -10,7 +10,7 @@ import typer
 
 from ..config.epresolution import NetworkType
 from ..sync.output import print_human_prune_results
-from ..sync.pruner import prune_all_syncs
+from ..snapshots.pruner import prune_all_syncs
 from .app import app
 from .common import (
     OutputFormat,
