+++ nbkp/snapshots/hardlinks.py
@@ -1,34 +1,38 @@
-"""Btrfs snapshot creation, lookup, and pruning."""
+"""Hard-link snapshot creation, lookup, symlink management, and pruning."""
 
 from __future__ import annotations
 
+import shutil
 from datetime import datetime, timezone
 
-from ...config import (
+from ..config import (
     Config,
+    LocalVolume,
+    RemoteVolume,
     SyncConfig,
     Volume,
 )
-from ...config.epresolution import ResolvedEndpoints
-from ...fsprotocol import SNAPSHOTS_DIR, STAGING_DIR
-from ...remote.dispatch import run_on_volume
+from ..config.epresolution import ResolvedEndpoints
+from ..fsprotocol import SNAPSHOTS_DIR
+from ..remote.dispatch import run_on_volume
 from .common import (
     create_snapshot_timestamp,
     list_snapshots,
+    read_latest_symlink,
     resolve_dest_path,
 )
 
 
-def create_snapshot(
+def create_snapshot_dir(
     sync: SyncConfig,
     config: Config,
     *,
     now: datetime | None = None,
     resolved_endpoints: ResolvedEndpoints | None = None,
 ) -> str:
-    """Create a read-only btrfs snapshot of staging/ into snapshots/.
+    """Create a snapshot directory for the current sync.
 
-    Returns the snapshot path.
+    Returns the full snapshot path.
     """
     re = resolved_endpoints or {}
     ts = now or datetime.now(timezone.utc)
@@ -37,31 +41,43 @@ def create_snapshot(
     dest_path = resolve_dest_path(sync, config)
     snapshot = create_snapshot_timestamp(ts, dst_vol)
     snapshot_path = f"{dest_path}/{SNAPSHOTS_DIR}/{snapshot.name}"
-    tmp_path = f"{dest_path}/{STAGING_DIR}"
-    result = run_on_volume(
-        ["btrfs", "subvolume", "snapshot", "-r", tmp_path, snapshot_path],
-        dst_vol,
-        re,
-    )
+    result = run_on_volume(["mkdir", "-p", snapshot_path], dst_vol, re)
+
     if result.returncode != 0:
-        raise RuntimeError(f"btrfs snapshot failed: {result.stderr}")
+        raise RuntimeError(f"mkdir snapshot dir failed: {result.stderr}")
     else:
         return snapshot_path
 
 
-def _make_snapshot_writable(
-    path: str,
-    volume: Volume,
-    resolved_endpoints: ResolvedEndpoints,
-) -> None:
-    """Unset the readonly property so the snapshot can be deleted."""
-    result = run_on_volume(
-        ["btrfs", "property", "set", path, "ro", "false"],
-        volume,
-        resolved_endpoints,
-    )
-    if result.returncode != 0:
-        raise RuntimeError(f"btrfs property set ro=false failed: {result.stderr}")
+def cleanup_orphaned_snapshots(
+    sync: SyncConfig,
+    config: Config,
+    *,
+    resolved_endpoints: ResolvedEndpoints | None = None,
+) -> list[str]:
+    """Remove snapshots newer than the latest symlink target.
+
+    These are leftover directories from failed syncs.
+    Returns list of deleted paths.
+    """
+    re = resolved_endpoints or {}
+    latest = read_latest_symlink(sync, config, resolved_endpoints=re)
+    if latest is None:
+        return []
+    else:
+        all_snapshots = list_snapshots(sync, config, re)
+        dst = config.destination_endpoint(sync)
+        dst_vol = config.volumes[dst.volume]
+        dest_path = resolve_dest_path(sync, config)
+        snapshots_dir = f"{dest_path}/{SNAPSHOTS_DIR}"
+        orphans = [
+            f"{snapshots_dir}/{s.name}"
+            for s in all_snapshots
+            if s.timestamp > latest.timestamp
+        ]
+        for path in orphans:
+            delete_snapshot(path, dst_vol, re)
+        return orphans
 
 
 def delete_snapshot(
@@ -69,20 +85,14 @@ def delete_snapshot(
     volume: Volume,
     resolved_endpoints: ResolvedEndpoints,
 ) -> None:
-    """Delete a single btrfs snapshot subvolume.
-
-    First unsets the readonly property (needed when the filesystem
-    is mounted with user_subvol_rm_allowed instead of granting
-    CAP_SYS_ADMIN), then deletes the subvolume.
-    """
-    _make_snapshot_writable(path, volume, resolved_endpoints)
-    result = run_on_volume(
-        ["btrfs", "subvolume", "delete", path],
-        volume,
-        resolved_endpoints,
-    )
-    if result.returncode != 0:
-        raise RuntimeError(f"btrfs delete failed: {result.stderr}")
+    """Delete a hard-link snapshot directory."""
+    match volume:
+        case RemoteVolume():
+            result = run_on_volume(["rm", "-rf", path], volume, resolved_endpoints)
+            if result.returncode != 0:
+                raise RuntimeError(f"rm -rf snapshot failed: {result.stderr}")
+        case LocalVolume():
+            shutil.rmtree(path)
 
 
 def prune_snapshots(
@@ -98,8 +108,6 @@ def prune_snapshots(
     Never prunes the snapshot that the latest symlink points to.
     Returns list of deleted (or would-be-deleted) paths.
     """
-    from .common import read_latest_symlink
-
     re = resolved_endpoints or {}
     snapshots = list_snapshots(sync, config, re)
     excess = len(snapshots) - max_snapshots
