+++ nbkp/sync/__init__.py
@@ -7,21 +7,18 @@ from typing import TYPE_CHECKING
 from .rsync import ProgressMode as ProgressMode
 
 if TYPE_CHECKING:
-    from .runner import PruneResult as PruneResult
     from .runner import SyncOutcome as SyncOutcome
     from .runner import SyncResult as SyncResult
     from .runner import run_all_syncs as run_all_syncs
 
 __all__ = [
     "ProgressMode",
-    "PruneResult",
     "SyncOutcome",
     "SyncResult",
     "run_all_syncs",
 ]
 
 _LAZY_MODULES = {
-    "PruneResult": "runner",
     "SyncOutcome": "runner",
     "SyncResult": "runner",
     "run_all_syncs": "runner",
