+++ tests/integration_docker/test_sync_hardlinks.py
@@ -17,12 +17,12 @@ from nbkp.config import (
 )
 from nbkp.config.epresolution import ResolvedEndpoints
 from nbkp.remote.resolution import resolve_all_endpoints
-from nbkp.sync.snapshots.common import (
+from nbkp.snapshots.common import (
     list_snapshots,
     read_latest_symlink,
     update_latest_symlink,
 )
-from nbkp.sync.snapshots.hardlinks import (
+from nbkp.snapshots.hardlinks import (
     cleanup_orphaned_snapshots,
     create_snapshot_dir,
     prune_snapshots,
