+++ tests/preflight/test_checks.py
@@ -21,7 +21,7 @@ from nbkp.config.epresolution import ResolvedEndpoint, ResolvedEndpoints
 from nbkp.remote.resolution import resolve_proxy_chain
 from nbkp.cli.common import OutputFormat
 from nbkp.sync import SyncResult
-from nbkp.sync.snapshots.common import create_snapshot_timestamp
+from nbkp.snapshots.common import create_snapshot_timestamp
 from nbkp.preflight.checks import check_all_syncs, check_sync
 from nbkp.preflight.queries import (
     _check_command_available,
